moved ui and auth dir inside interal dir
