+++ internal/ui/ui.go
@@ -1,0 +1,1 @@
+package ui
